SCST app starters GA release configuration changes

+++ src/main/groovy/org/springframework/jenkins/scstappstarters/common/SpringScstAppStarterJobs.groovy
@@ -60,8 +60,23 @@ trait SpringScstAppStarterJobs extends BuildAndDeploy {
 		return "rm -rf /tmp/gitcredentials"
 	}
 
-	String cleanAndDeploy(String releaseVersion) {
-		return """
+	String cleanAndDeploy(boolean isRelease, String releaseType) {
+		if (isRelease && releaseType != null && !releaseType.equals("milestone")) {
+
+			return  """
+                        lines=\$(find . -type f -name pom.xml | xargs egrep "SNAPSHOT|M[0-9]|RC[0-9]" | grep -v regex | wc -l)
+                        if [ \$lines -eq 0 ]; then
+                            set +x
+                            ./mvnw clean deploy -Pspring -Dgpg.secretKeyring="\$${gpgSecRing()}" -Dgpg.publicKeyring="\$${
+				gpgPubRing()}" -Dgpg.passphrase="\$${gpgPassphrase()}" -DSONATYPE_USER="\$${sonatypeUser()}" -DSONATYPE_PASSWORD="\$${sonatypePassword()}" -Pcentral -U
+                            set -x
+                        else
+                            echo "Non release versions found. Aborting build"
+                        fi
+                    """
+		}
+		if (isRelease && releaseType != null && releaseType.equals("milestone")) {
+			return """
 					#!/bin/bash -x
 
 			   		lines=\$(find . -type f -name pom.xml | xargs grep SNAPSHOT | wc -l)
@@ -71,21 +86,39 @@ trait SpringScstAppStarterJobs extends BuildAndDeploy {
 						echo "Snapshots found. Aborting the release build."
 					fi
 			   """
-
+		}
 	}
 
-	String cleanAndDeployWithGenerateApps() {
-		return """
-					#!/bin/bash -x
-					rm -rf apps
+	String cleanAndDeployWithGenerateApps(boolean isRelease, String releaseType) {
+		if (isRelease && releaseType != null && !releaseType.equals("milestone")) {
+            return """
+                    #!/bin/bash -x
+                    rm -rf apps
 
-					lines=\$(find . -type f -name pom.xml | xargs grep SNAPSHOT | wc -l)
-					if [ \$lines -eq 0 ]; then
-						./mvnw clean deploy -U -Pspring -PgenerateApps
-					else
-						echo "Snapshots found. Aborting the release build."
-					fi
-			   """
+                    lines=\$(find . -type f -name pom.xml | xargs egrep "SNAPSHOT|M[0-9]|RC[0-9]" | grep -v regex | wc -l)
+                    if [ \$lines -eq 0 ]; then
+                        set +x
+                        ./mvnw clean deploy -Pspring -PgenerateApps -Dgpg.secretKeyring="\$${gpgSecRing()}" -Dgpg.publicKeyring="\$${
+                gpgPubRing()}" -Dgpg.passphrase="\$${gpgPassphrase()}" -DSONATYPE_USER="\$${sonatypeUser()}" -DSONATYPE_PASSWORD="\$${sonatypePassword()}" -Pcentral -U
+                        set -x
+                    else
+                        echo "Non release versions found. Aborting build"
+                    fi
+                """
+        }
+        if (isRelease && releaseType != null && releaseType.equals("milestone")) {
+            return """
+                #!/bin/bash -x
+                rm -rf apps
+
+                lines=\$(find . -type f -name pom.xml | xargs grep SNAPSHOT | wc -l)
+                if [ \$lines -eq 0 ]; then
+                    ./mvnw clean deploy -U -Pspring -PgenerateApps
+                else
+                    echo "Snapshots found. Aborting the release build."
+                fi
+           """
+        }
 	}
 
 	String removeAppsDirectory() {
@@ -95,9 +128,25 @@ trait SpringScstAppStarterJobs extends BuildAndDeploy {
 			   """
 	}
 
-	String cleanAndInstall() {
+	String cleanAndInstall(boolean isRelease, String releaseType) {
+        if (isRelease && releaseType != null && !releaseType.equals("milestone")) {
+            return """
+                    #!/bin/bash -x
+                    rm -rf apps
 
-		return """
+                    lines=\$(find . -type f -name pom.xml | xargs egrep "SNAPSHOT|M[0-9]|RC[0-9]" | grep -v regex | wc -l)
+                    if [ \$lines -eq 0 ]; then
+                        set +x
+                        ./mvnw clean deploy -Pspring -Dgpg.secretKeyring="\$${gpgSecRing()}" -Dgpg.publicKeyring="\$${
+                gpgPubRing()}" -Dgpg.passphrase="\$${gpgPassphrase()}" -DSONATYPE_USER="\$${sonatypeUser()}" -DSONATYPE_PASSWORD="\$${sonatypePassword()}" -Pcentral -U
+                        set -x
+                    else
+                        echo "Non release versions found. Aborting build"
+                    fi
+                """
+        }
+        if (isRelease && releaseType != null && releaseType.equals("milestone")) {
+            return """
 					#!/bin/bash -x
 
 			   		lines=\$(find . -type f -name pom.xml | xargs grep SNAPSHOT | wc -l)
@@ -107,5 +156,26 @@ trait SpringScstAppStarterJobs extends BuildAndDeploy {
 						echo "Snapshots found. Aborting the release build."
 					fi
 			   """
+        }
+	}
+
+	String gpgSecRing() {
+		return 'FOO_SEC'
+	}
+
+	String gpgPubRing() {
+		return 'FOO_PUB'
+	}
+
+	String gpgPassphrase() {
+		return 'FOO_PASSPHRASE'
+	}
+
+	String sonatypeUser() {
+		return 'SONATYPE_USER'
+	}
+
+	String sonatypePassword() {
+		return 'SONATYPE_PASSWORD'
 	}
 }
